pyproject.toml
[tool.poetry]
name = "easy-check"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.8"}
sqlmodel = "^0.0.22"
sqlalchemy-utils = "^0.41.2"
passlib = "^1.7.4"
pydantic-settings = "^2.7.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
isort = "^6.0.0"
flake8 = "^7.1.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
